BackendAssesmentCopy/backend/assesment/requirements.txt
fastapi==0.63.0
cython==0.29.22
pydantic==1.8
sqlalchemy==1.3.23
uvicorn==0.13.4
async-exit-stack==1.0.1
async-generator==1.10
psycopg2==2.8.6
kafka-python==2.0.2